chore([].find): Don't touch Array.prototype
Even if the provided polyfill for Array.find was valid and behaved
according to the ES6 spec, I removed it from the code. It is bad practice
to mess with core prototype's.

This change was triggered by a problem in combination with another library
which indeed replaced Array.find, too, but not according to the ES6 spec.