feat(resolver): default resolver now accepts require
It is now possible to pass in an external require to the default resolver.
This is needed, if tiny-di is not installed as node_module of the application.
This is e.g. the case, if it is linked via `npm link`:

```bash
/some/tiny-di: npm link
/some/app: npm link tiny-di
```

The real path of tiny-di in this scenario is /some/tiny-di. Requiring modules
which are installed under /some/app/node_modules will fail.

To solve this, the app should do this:

```js
var $injector = new Injector();
$injector.setResolve($injector.getResolver(require));
```

Doing this will pass the require of the app to tiny-di. The require of the
app can load modules from /some/app/node_modules